Jcf/stable card loading (#942)
* Optimistically list element card

* Skeleton rects for loading card content

* Stablising loading for market and batch

* Inject assets from initial status fetch allowing
recipes to be hydrated before sdk initialisation.
This also makes many observables redundant.

* - remove dead config field: maxAvailableAssetId
- reduce tx limits

+++ zk-money/src/alt-model/price_feeds/price_fetchers.ts
@@ -23,7 +23,7 @@ function createDefaultPriceFetcher(priceFeedContractAddressStr: string, provider
   return async () => {
     try {
       const bigNum = await contract.latestAnswer();
-      return bigNum.toBigInt();
+      return bigNum.toBigInt() as bigint;
     } catch (err) {
       debug(`Price fetch failed for address ${priceFeedContractAddressStr}`, err);
       throw err;
